+++ package.json
@@ -1,20 +1,17 @@
 {
   "name": "typedoc-plugin-rename-defaults",
-  "version": "0.6.6",
+  "version": "0.6.7",
+  "type": "module",
   "description": "A TypeDoc plugin that renames the `default` exports to their original name",
-  "main": "dist/index.js",
-  "files": [],
+  "exports": "./index.js",
+  "files": [
+    "index.js"
+  ],
   "scripts": {
-    "build": "tsc",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "test:manual": "npm run build && typedoc && http-server docs -c-1 --silent -o",
-    "test:manual-no-plugin": "npm run build && typedoc && http-server docs -c-1 --silent -o",
-    "prepack": "rm -rf dist && npm run build"
+    "dev": "cd test && typedoc && serve docs",
+    "format": "prettier -w ."
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/felipecrs/typedoc-plugin-rename-defaults.git"
-  },
+  "repository": "github:felipecrs/typedoc-plugin-rename-defaults",
   "keywords": [
     "typedoc",
     "plugin",
@@ -23,21 +20,15 @@
   ],
   "author": "Felipe Santos (https://github.com/felipecrs)",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/felipecrs/typedoc-plugin-rename-defaults/issues"
-  },
+  "bugs": "https://github.com/felipecrs/typedoc-plugin-rename-defaults/issues",
   "homepage": "https://github.com/felipecrs/typedoc-plugin-rename-defaults#readme",
   "peerDependencies": {
     "typedoc": "0.22.x || 0.23.x || 0.24.x || 0.25.x"
   },
   "devDependencies": {
-    "http-server": "^14.1.1",
-    "typedoc": "^0.25.0",
-    "typedoc-plugin-rename-defaults": ".",
-    "typescript": "^5.0.4"
-  },
-  "volta": {
-    "node": "18.17.1",
-    "npm": "9.8.1"
+    "prettier": "^3.0.3",
+    "serve": "^14.2.1",
+    "typedoc": "^0.25.2",
+    "typescript": "^5.2.2"
   }
 }
