--- tsconfig.typedoc.json
@@ -1,6 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "include": [
-        "test"
-    ],
-}
