--- typedoc.json
@@ -1,7 +1,0 @@
-{
-    "$schema": "https://typedoc.org/schema.json",
-    "entryPointStrategy": "Expand",
-    "readme": "none",
-    "entryPoints": ["test"],
-    "tsconfig": "./tsconfig.typedoc.json"
-}
